requirements.txt
pandas
numpy
yfinance
fredapi
python-dotenv
pyarrow
fastparquet
scikit-learn